+++ src/expression.cpp
@@ -14,6 +14,7 @@
 #include <cstddef>
 #include <cstdint>
 #include <deque>
+#include <exception>
 #include <functional>
 #include <iterator>
 #include <limits>
@@ -22,6 +23,7 @@
 #include <optional>
 #include <ostream>
 #include <set>
+#include <sstream>
 #include <stdexcept>
 #include <string>
 #include <tuple>
@@ -38,6 +40,7 @@
 #include <tbb/blocked_range.h>
 #include <tbb/parallel_for.h>
 
+#include <llvm/Config/llvm-config.h>
 #include <llvm/IR/Attributes.h>
 #include <llvm/IR/BasicBlock.h>
 #include <llvm/IR/Constant.h>
@@ -70,6 +73,7 @@
 #endif
 
 #include <heyoka/detail/cm_utils.hpp>
+#include <heyoka/detail/func_cache.hpp>
 #include <heyoka/detail/fwd_decl.hpp>
 #include <heyoka/detail/llvm_fwd.hpp>
 #include <heyoka/detail/llvm_helpers.hpp>
@@ -161,7 +165,7 @@ namespace detail
 namespace
 {
 
-expression copy_impl(std::unordered_map<const void *, expression> &func_map, const expression &e)
+expression copy_impl(funcptr_map<expression> &func_map, const expression &e)
 {
     return std::visit(
         [&func_map](const auto &v) {
@@ -212,7 +216,8 @@ expression copy_impl(std::unordered_map<const void *, expression> &func_map, con
 
 expression copy(const expression &e)
 {
-    std::unordered_map<const void *, expression> func_map;
+    thread_local detail::funcptr_map<expression> func_map;
+    func_map.clear();
 
     return detail::copy_impl(func_map, e);
 }
@@ -299,7 +304,7 @@ namespace detail
 namespace
 {
 
-void get_variables(std::unordered_set<const void *> &func_set, std::set<std::string> &s_set, const expression &e)
+void get_variables(funcptr_set &func_set, std::set<std::string> &s_set, const expression &e)
 {
     std::visit(
         [&func_set, &s_set](const auto &arg) {
@@ -331,7 +336,7 @@ void get_variables(std::unordered_set<const void *> &func_set, std::set<std::str
         e.value());
 }
 
-void rename_variables(std::unordered_set<const void *> &func_set, expression &e,
+void rename_variables(funcptr_set &func_set, expression &e,
                       const std::unordered_map<std::string, std::string> &repl_map)
 {
     std::visit(
@@ -370,17 +375,24 @@ void rename_variables(std::unordered_set<const void *> &func_set, expression &e,
 
 std::vector<std::string> get_variables(const expression &e)
 {
-    std::unordered_set<const void *> func_set;
-    std::set<std::string> s_set;
+    thread_local detail::funcptr_set func_set;
+    func_set.clear();
+
+    // NOTE: it's probably better here to eventually
+    // change this to a fast unordered set, copy the items
+    // into a vector and sort it on output.
+    thread_local std::set<std::string> s_set;
+    s_set.clear();
 
     detail::get_variables(func_set, s_set, e);
 
-    return std::vector<std::string>(s_set.begin(), s_set.end());
+    return {s_set.begin(), s_set.end()};
 }
 
 void rename_variables(expression &e, const std::unordered_map<std::string, std::string> &repl_map)
 {
-    std::unordered_set<const void *> func_set;
+    thread_local detail::funcptr_set func_set;
+    func_set.clear();
 
     detail::rename_variables(func_set, e, repl_map);
 }
@@ -390,20 +402,99 @@ void swap(expression &ex0, expression &ex1) noexcept
     std::swap(ex0.value(), ex1.value());
 }
 
-// NOTE: this implementation does not take advantage of potentially
-// repeating subexpressions. This is not currently a problem because
-// hashing is needed only in the CSE for the decomposition, which involves
-// only trivial expressions. However, this would likely be needed by a to_sympy()
-// implementation in heyoka.py which allows for a dictionary of custom
-// substitutions to be provided by the user.
+namespace detail
+{
+
+std::size_t hash(funcptr_map<std::size_t> &func_map, const expression &ex)
+{
+    return std::visit(
+        [&func_map](const auto &v) {
+            using type = detail::uncvref_t<decltype(v)>;
+
+            if constexpr (std::is_same_v<type, func>) {
+                const auto f_id = v.get_ptr();
+
+                if (auto it = func_map.find(f_id); it != func_map.end()) {
+                    // We already computed the hash for the current
+                    // function, return it.
+                    return it->second;
+                }
+
+                // Compute the hash of the current function.
+                auto retval = v.hash(func_map);
+
+                // Add it to the cache.
+                [[maybe_unused]] const auto [_, flag] = func_map.emplace(f_id, retval);
+                // NOTE: an expression cannot contain itself.
+                assert(flag);
+
+                return retval;
+            } else {
+                return hash(v);
+            }
+        },
+        ex.value());
+}
+
+} // namespace detail
+
 std::size_t hash(const expression &ex)
 {
-    return std::visit([](const auto &v) { return hash(v); }, ex.value());
+    thread_local detail::funcptr_map<std::size_t> func_map;
+    func_map.clear();
+
+    return detail::hash(func_map, ex);
 }
+
+namespace detail
+{
+
+namespace
+{
+
+// Exception to signal that the stream output
+// for an expression has become too large.
+struct output_too_long : std::exception {
+};
+
+} // namespace
+
+// Helper to stream an expression to a stringstream, while
+// checking that the number of characters written so far
+// to the stream is not too large. If that is the case,
+// an exception will be thrown.
+void stream_expression(std::ostringstream &oss, const expression &e)
+{
+    if (oss.tellp() > 1000) {
+        throw output_too_long{};
+    }
+
+    std::visit(
+        [&oss](const auto &v) {
+            using type = detail::uncvref_t<decltype(v)>;
+
+            if constexpr (std::is_same_v<type, func>) {
+                v.to_stream(oss);
+            } else {
+                oss << v;
+            }
+        },
+        e.value());
+}
+
+} // namespace detail
 
 std::ostream &operator<<(std::ostream &os, const expression &e)
 {
-    return std::visit([&os](const auto &arg) -> std::ostream & { return os << arg; }, e.value());
+    std::ostringstream oss;
+
+    try {
+        detail::stream_expression(oss, e);
+    } catch (const detail::output_too_long &) {
+        oss << "...";
+    }
+
+    return os << oss.str();
 }
 
 expression operator+(expression e)
@@ -1027,7 +1118,7 @@ namespace detail
 namespace
 {
 
-std::size_t get_n_nodes(std::unordered_map<const void *, std::size_t> &func_map, const expression &e)
+std::size_t get_n_nodes(funcptr_map<std::size_t> &func_map, const expression &e)
 {
     return std::visit(
         [&func_map](const auto &arg) -> std::size_t {
@@ -1065,7 +1156,8 @@ std::size_t get_n_nodes(std::unordered_map<const void *, std::size_t> &func_map,
 
 std::size_t get_n_nodes(const expression &e)
 {
-    std::unordered_map<const void *, std::size_t> func_map;
+    thread_local detail::funcptr_map<std::size_t> func_map;
+    func_map.clear();
 
     return detail::get_n_nodes(func_map, e);
 }
@@ -1073,7 +1165,7 @@ std::size_t get_n_nodes(const expression &e)
 namespace detail
 {
 
-expression diff(std::unordered_map<const void *, expression> &func_map, const expression &e, const std::string &s)
+expression diff(funcptr_map<expression> &func_map, const expression &e, const std::string &s)
 {
     return std::visit(
         [&func_map, &s](const auto &arg) {
@@ -1118,7 +1210,7 @@ expression diff(std::unordered_map<const void *, expression> &func_map, const ex
         e.value());
 }
 
-expression diff(std::unordered_map<const void *, expression> &func_map, const expression &e, const param &p)
+expression diff(funcptr_map<expression> &func_map, const expression &e, const param &p)
 {
     return std::visit(
         [&func_map, &p](const auto &arg) {
@@ -1162,14 +1254,16 @@ expression diff(std::unordered_map<const void *, expression> &func_map, const ex
 
 expression diff(const expression &e, const std::string &s)
 {
-    std::unordered_map<const void *, expression> func_map;
+    thread_local detail::funcptr_map<expression> func_map;
+    func_map.clear();
 
     return detail::diff(func_map, e, s);
 }
 
 expression diff(const expression &e, const param &p)
 {
-    std::unordered_map<const void *, expression> func_map;
+    thread_local detail::funcptr_map<expression> func_map;
+    func_map.clear();
 
     return detail::diff(func_map, e, p);
 }
@@ -1196,7 +1290,7 @@ namespace detail
 namespace
 {
 
-expression subs(std::unordered_map<const void *, expression> &func_map, const expression &ex,
+expression subs(funcptr_map<expression> &func_map, const expression &ex,
                 const std::unordered_map<std::string, expression> &smap)
 {
     return std::visit(
@@ -1246,7 +1340,8 @@ expression subs(std::unordered_map<const void *, expression> &func_map, const ex
 
 expression subs(const expression &e, const std::unordered_map<std::string, expression> &smap)
 {
-    std::unordered_map<const void *, expression> func_map;
+    thread_local detail::funcptr_map<expression> func_map;
+    func_map.clear();
 
     return detail::subs(func_map, e, smap);
 }
@@ -1461,8 +1556,8 @@ void update_grad_dbl(std::unordered_map<std::string, double> &grad, const expres
 namespace detail
 {
 
-taylor_dc_t::size_type taylor_decompose(std::unordered_map<const void *, taylor_dc_t::size_type> &func_map,
-                                        const expression &ex, taylor_dc_t &dc)
+taylor_dc_t::size_type taylor_decompose(funcptr_map<taylor_dc_t::size_type> &func_map, const expression &ex,
+                                        taylor_dc_t &dc)
 {
     if (const auto *fptr = std::get_if<func>(&ex.value())) {
         return fptr->taylor_decompose(func_map, dc);
@@ -1478,7 +1573,8 @@ taylor_dc_t::size_type taylor_decompose(std::unordered_map<const void *, taylor_
 // If the return value is zero, ex was not decomposed.
 taylor_dc_t::size_type taylor_decompose(const expression &ex, taylor_dc_t &dc)
 {
-    std::unordered_map<const void *, taylor_dc_t::size_type> func_map;
+    thread_local detail::funcptr_map<taylor_dc_t::size_type> func_map;
+    func_map.clear();
 
     return detail::taylor_decompose(func_map, ex, dc);
 }
@@ -1587,7 +1683,7 @@ namespace detail
 namespace
 {
 
-std::uint32_t get_param_size(std::unordered_set<const void *> &func_set, const expression &ex)
+std::uint32_t get_param_size(detail::funcptr_set &func_set, const expression &ex)
 {
     std::uint32_t retval = 0;
 
@@ -1634,7 +1730,8 @@ std::uint32_t get_param_size(std::unordered_set<const void *> &func_set, const e
 // is zero, no params appear in the expression.
 std::uint32_t get_param_size(const expression &ex)
 {
-    std::unordered_set<const void *> func_set;
+    thread_local detail::funcptr_set func_set;
+    func_set.clear();
 
     return detail::get_param_size(func_set, ex);
 }
@@ -1645,7 +1742,7 @@ namespace detail
 namespace
 {
 
-bool is_time_dependent(std::unordered_map<const void *, bool> &func_map, const expression &ex)
+bool is_time_dependent(funcptr_map<bool> &func_map, const expression &ex)
 {
     // - If ex is a function, check if it is time-dependent, or
     //   if any of its arguments is time-dependent,
@@ -1717,9 +1814,8 @@ std::vector<std::pair<expression, expression>> copy(const std::vector<std::pair<
     return ret;
 }
 
-std::optional<std::vector<expression>::size_type>
-decompose(std::unordered_map<const void *, std::vector<expression>::size_type> &func_map, const expression &ex,
-          std::vector<expression> &dc)
+std::optional<std::vector<expression>::size_type> decompose(funcptr_map<std::vector<expression>::size_type> &func_map,
+                                                            const expression &ex, std::vector<expression> &dc)
 {
     if (const auto *fptr = std::get_if<func>(&ex.value())) {
         return fptr->decompose(func_map, dc);
@@ -2105,7 +2201,8 @@ std::vector<expression> function_sort_dc(std::vector<expression> &dc, std::vecto
 
 std::optional<std::vector<expression>::size_type> decompose(const expression &ex, std::vector<expression> &dc)
 {
-    std::unordered_map<const void *, std::vector<expression>::size_type> func_map;
+    thread_local detail::funcptr_map<std::vector<expression>::size_type> func_map;
+    func_map.clear();
 
     return detail::decompose(func_map, ex, dc);
 }
@@ -3285,7 +3382,19 @@ auto add_cfunc_impl(llvm_state &s, const std::string &name, const F &fn, std::ui
     builder.SetInsertPoint(bb);
 
     // Invoke the strided function with stride == batch_size.
-    builder.CreateCall(f_strided, {out_ptr, in_ptr, par_ptr, time_ptr, to_size_t(s, builder.getInt32(batch_size))});
+    auto *fcall = builder.CreateCall(f_strided,
+                                     {out_ptr, in_ptr, par_ptr, time_ptr, to_size_t(s, builder.getInt32(batch_size))});
+    // NOTE: forcibly inline the function call. This will increase
+    // compile time, but we want to make sure that the non-strided
+    // version of the compiled function is optimised as much as possible,
+    // so we accept the tradeoff.
+#if LLVM_VERSION_MAJOR >= 14
+    fcall->addFnAttr(llvm::Attribute::AlwaysInline);
+#else
+    auto attrs = fcall->getAttributes();
+    attrs = attrs.addAttribute(context, llvm::AttributeList::FunctionIndex, llvm::Attribute::AlwaysInline);
+    fcall->setAttributes(attrs);
+#endif
 
     // Finish off the function.
     builder.CreateRetVoid();
@@ -3367,7 +3476,8 @@ template HEYOKA_DLL_PUBLIC std::vector<expression> add_cfunc<mppp::real>(llvm_st
 // Determine if an expression is time-dependent.
 bool is_time_dependent(const expression &ex)
 {
-    std::unordered_map<const void *, bool> func_map;
+    thread_local detail::funcptr_map<bool> func_map;
+    func_map.clear();
 
     return detail::is_time_dependent(func_map, ex);
 }