+++ src/math/binary_op.cpp
@@ -14,7 +14,7 @@
 #include <cstdint>
 #include <functional>
 #include <initializer_list>
-#include <ostream>
+#include <sstream>
 #include <stdexcept>
 #include <string>
 #include <type_traits>
@@ -40,6 +40,7 @@
 
 #endif
 
+#include <heyoka/detail/func_cache.hpp>
 #include <heyoka/detail/fwd_decl.hpp>
 #include <heyoka/detail/llvm_helpers.hpp>
 #include <heyoka/detail/string_conv.hpp>
@@ -79,29 +80,33 @@ std::size_t binary_op::extra_hash() const
     return std::hash<type>{}(m_type);
 }
 
-void binary_op::to_stream(std::ostream &os) const
+void binary_op::to_stream(std::ostringstream &oss) const
 {
     assert(args().size() == 2u);
     assert(m_type >= type::add && m_type <= type::div);
 
-    os << '(' << lhs() << ' ';
+    oss << '(';
+    stream_expression(oss, lhs());
+    oss << ' ';
 
     switch (m_type) {
         case type::add:
-            os << '+';
+            oss << '+';
             break;
         case type::sub:
-            os << '-';
+            oss << '-';
             break;
         case type::mul:
-            os << '*';
+            oss << '*';
             break;
         default:
-            os << '/';
+            oss << '/';
             break;
     }
 
-    os << ' ' << rhs() << ')';
+    oss << ' ';
+    stream_expression(oss, rhs());
+    oss << ')';
 }
 
 binary_op::type binary_op::op() const
@@ -122,7 +127,7 @@ const expression &binary_op::rhs() const
 }
 
 template <typename T>
-expression binary_op::diff_impl(std::unordered_map<const void *, expression> &func_map, const T &x) const
+expression binary_op::diff_impl(funcptr_map<expression> &func_map, const T &x) const
 {
     assert(args().size() == 2u);
     assert(m_type >= type::add && m_type <= type::div);
@@ -140,12 +145,12 @@ expression binary_op::diff_impl(std::unordered_map<const void *, expression> &fu
     }
 }
 
-expression binary_op::diff(std::unordered_map<const void *, expression> &func_map, const std::string &s) const
+expression binary_op::diff(funcptr_map<expression> &func_map, const std::string &s) const
 {
     return diff_impl(func_map, s);
 }
 
-expression binary_op::diff(std::unordered_map<const void *, expression> &func_map, const param &p) const
+expression binary_op::diff(funcptr_map<expression> &func_map, const param &p) const
 {
     return diff_impl(func_map, p);
 }